ai-service-vision/pyproject.toml
[project]
name = "ai-service-vision"
version = "0.1.0"
description = "YOLO-based vision service for iPhone model detection and damage analysis"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pillow>=11.0.0",
    "httpx>=0.27.0",
    "ultralytics>=8.3.0",
    "numpy>=1.26.0,<2.0.0",
    "opencv-python-headless>=4.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "ipykernel>=6.29.0",
]
scrape = [
    "playwright>=1.48.0",
    "beautifulsoup4>=4.12.0",
    "tqdm>=4.66.0",
    "google-generativeai>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]
